IMPORTANT: Router now bubbles error events from commands.
Before, an error emitted from a command might not be handled
by the router, which means that Node.js throws it. And this means
that anything that catches errors might catch this event (and
then mishandle it). So now the router catches and emits, which
means the outter error handler (HTTP server, CLI, whatever) can
now react to a command error.